// packages/helper/src/components/tag-input/tag-input-types.ts
export type TagInputSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl'

export type TagInputColor =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'teal'
  | 'purple'
  | 'neutral'
  | 'dark'

export type TagInputVariant = 'outlined' | 'filled' | 'underlined'
// ...
export type TagInputValidationState =
  | 'default'
  | 'valid'
  | 'invalid'
  | 'warning'

export type TagInputValue = unknown

export interface TagInputItem {
  label?: string
  value?: unknown
  description?: string
  disabled?: boolean
  [key: string]: unknown
}